+++ public/scripts/app.js
@@ -21,7 +21,7 @@ import pcr from './plugins/pcr/plugin';
 import sequencing_primers from './plugins/sequencing_primers/plugin';
 import chromatograph from './plugins/chromatograph/plugin';
 
-import tooltip from 'gentle-utils/tooltip';
+import tooltip from 'tooltip';
 
 var plugins = [ncbi, designer, blast, pcr, sequencing_primers, chromatograph];