+++ public/scripts/library/rdp/rdp_sequence_transform.js
@@ -2,9 +2,10 @@ import _ from 'underscore';
 import RdpEdit from './rdp_edit';
 import {
   firstCodonIsMethionine,
-  noTerminalStopCodons,
+  lastStopCodonsRemoved,
   ensureLastBaseIs,
   firstCodonIsStop,
+  lastCodonIsStop,
   warnIfEarlyStopCodons,
 } from './sequence_transform';
 
@@ -18,10 +19,15 @@ var calculateTransformationFunctionInstances = function(sequenceModel) {
     if(desiredStickyEnds.start.name === 'X') {
       transforms.push(firstCodonIsMethionine);
     }
-    transforms.push(noTerminalStopCodons);
 
-    var lastBaseMustBe = desiredStickyEnds.end.sequence.substr(0, 1);
-    transforms.push(ensureLastBaseIs(lastBaseMustBe));
+    if(sequenceModel.isCdsWithStop) {
+      transforms.push(lastCodonIsStop);
+    } else {
+      transforms.push(lastStopCodonsRemoved);
+
+      var lastBaseMustBe = desiredStickyEnds.end.sequence.substr(0, 1);
+      transforms.push(ensureLastBaseIs(lastBaseMustBe));
+    }
 
     transforms.push(warnIfEarlyStopCodons);
   } else if(sequenceModel.isRBS || sequenceModel.isTerminator) {