+++ public/scripts/library/rdp/wip_rdp_oligo_sequence.js
@@ -9,13 +9,10 @@ var wip_rdp_oligo_sequence = 'wip_rdp_oligo_sequence';
 
 
 class WipRdpOligoSequence extends WipRdpAbstractSequence {
-  constructor(attrs, ...args) {
-    if(attrs._type && attrs._type !== wip_rdp_oligo_sequence) {
-      throw new TypeError(`WipRdpPcrSequence expected _type of "${wip_rdp_oligo_sequence}" but was "${attrs._type}"`);
-    }
-    attrs.Klass = WipRdpOligoSequence;
-    attrs.types = RdpTypes.oligoTypes;
-    super(attrs, ...args);
+  constructor(attrs, options={}) {
+    options.Klass = WipRdpOligoSequence;
+    options.types = RdpTypes.oligoTypes;
+    super(attrs, options);
     this.set({_type: wip_rdp_oligo_sequence}, {silent: true});
   }