+++ public/scripts/plugins/designer/lib/wip_circuit.js
@@ -287,7 +287,7 @@ export default class WipCircuit extends Sequence {
     
     var finalSequence = Sequence.concatenateSequences(
       sequences, 
-      isCircular
+      false
     );
 
     finalSequence.set({