+++ public/scripts/plugins/pcr/lib/product.js
@@ -1,4 +1,4 @@
-import {version1GenericPreProcessor} from 'gentle-utils/preprocessor';
+import {version1GenericPreProcessor} from 'gentledna-utils/dist/preprocessor';
 import SequenceModel from '../../../sequence/models/sequence';
 import SequencesCollection from '../../../sequence/models/sequences';
 import PcrPrimer from './pcr_primer';
@@ -13,6 +13,17 @@ class PcrProductSequence extends SequenceModel {
     this.set({_type: pcr_product}, {silent: true});
   }
 
+  defaults() {
+    var defaults = super.defaults();
+    return _.extend(super.defaults(), {
+      displaySettings: _.extend({}, defaults.displaySettings, {
+        rows: _.extend({}, defaults.displaySettings.rows || {}, {
+          aaOffset: 1,
+        })
+      })
+    });
+  }
+
   get requiredFields() {
     return super.requiredFields.concat([
       'forwardPrimer',
@@ -22,11 +33,12 @@ class PcrProductSequence extends SequenceModel {
   }
 
   validateFields(attributes) {
-    super.validateFields(attributes);
+    var errors = super.validateFields(attributes);
     // Remove once stickyEnds are a class in their own right and their
     // constructor would raise this error instead / return an empty stickyEnds
     // instance
-    if(!attributes.stickyEnds) throw new TypeError(`stickyEnds provided as attribute but were: ${attributes.stickyEnds}`);
+    if(!attributes.stickyEnds) errors.push(`stickyEnds provided as attribute but were: ${attributes.stickyEnds}`);
+    return errors;
   }
 }