+++ public/scripts/plugins/pcr/views/pcr_form_view.js
@@ -10,13 +10,13 @@ import WipRdpOligoSequence from 'gentle-rdp/wip_rdp_oligo_sequence';
 import template from '../templates/pcr_form_view.hbs';
 import rdpErrorsTemplate from '../templates/rdp_errors.hbs';
 import EditsView from './pcr_edits_view';
-import {humaniseRdpType} from '../lib/utils';
+import {humaniseRdpLabel} from '../lib/utils';
 
 
 var convertForSelect = function(values) {
   return _.map(values, (value) => {
     return {
-      name: humaniseRdpType(value),
+      name: humaniseRdpLabel(value),
       value
     };
   });
@@ -44,11 +44,20 @@ export default Backbone.View.extend({
     this.hasRdpOligoSequence = this.model instanceof WipRdpOligoSequence;
     this.hasRdpPcrSequence = !this.hasRdpOligoSequence;
     var partType = this.model.get('partType');
+
+    
+    // if description has not been set/modified, set it to name
+    var desc = this.model.get("desc")
+    if(!desc) {
+      desc = this.model.get('name')
+    }
+
     this.state = _.defaults({
       from: selectionFrom || 0,
       to: selectionTo || this.model.getLength(this.model.STICKY_END_ANY)-1,
       name: this.model.get('name'),
       sourceSequenceName: this.model.get('sourceSequenceName'),
+      desc: desc,
     }, this.model.get('meta.pcr.defaults') || {}, {
       targetMeltingTemperature: 68.5, 
       partType,
@@ -70,7 +79,14 @@ export default Backbone.View.extend({
   },
 
   availablePartTypes: function() {
-    return convertForSelect(this.model.availablePartTypes);
+    var baseTypes = convertForSelect(this.model.availablePartTypes);
+    var blank = {name: "Select RDP Part Type",
+                 value: "",
+                 disabled: true,
+                 isSelected: true};
+    
+    baseTypes.unshift(blank)
+    return baseTypes;
   },
 
   availableStickyEndNameOptions: function() {
@@ -192,13 +208,15 @@ export default Backbone.View.extend({
     try {
       if(this.state.invalid.any) {
         alert('Some RDP part details are incorrect or missing.  Please correct them first.');
+        console.log(this.state.invalid)
       } else {
-        var data = this.getData();
-        var attributes = _.pick(data, 'name', 'sequence', 'features', 'sourceSequenceName', 'partType', 'desiredStickyEnds');
-        attributes.frm = data.from;
-        attributes.size = data.to - data.from + 1;
+        var attributes = this.getData();
+        attributes.frm = attributes.from;
+        attributes.size = attributes.to - attributes.from + 1;
+        delete attributes.from;
+        delete attributes.to;
         // TODO refactor to keep same sequenceModel?
-        var desiredWipRdpSequence = this.model.getRdpCompliantSequenceModel(attributes);
+        var desiredWipRdpSequence = this.model.getWipRdpCompliantSequenceModel(attributes);
         var rdpEdits = desiredWipRdpSequence.get('rdpEdits');
         var errors = desiredWipRdpSequence.errors();
 
@@ -226,10 +244,10 @@ export default Backbone.View.extend({
 
   createNewRdpPart: function(desiredWipRdpSequence) {
     this.state.calculating = true;
-    var data = this.getData();
-    data.shortName = data.desiredStickyEnds.start.name +'-'+ data.shortName +'-'+ data.desiredStickyEnds.end.name;
     if(this.hasRdpOligoSequence) {
       var wipRdpOligoSequence = desiredWipRdpSequence;
+      var data = this.getData();
+      data.shortName = data.desiredStickyEnds.start.name +'-'+ data.shortName +'-'+ data.desiredStickyEnds.end.name;
       data.stickyEnds = data.desiredStickyEnds;
       var newRdpOligoSequence = wipRdpOligoSequence.getRdpOligoSequence(data);